requirements.txt
bitstring==3.1.5
pycryptodome==3.14.1